Move application stuff into applicationcontroller.
